requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cmake==3.25.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
datasets==3.6.0
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.0
fastcore==1.8.2
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
ghapi==1.0.6
gitdb==4.0.12
GitPython==3.1.44
grpclib==0.4.7
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
html2text==2024.2.26
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.1
hyperframe==6.1.0
identify==2.6.12
idna==3.10
javalang==0.13.0
Jinja2==3.1.4
jiter==0.10.0
kiwisolver==1.4.8
libclang==18.1.1
lit==15.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
modal==1.0.1
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
neo4j==5.28.1
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
openai==1.67.0
optimum==1.25.3
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylcs==0.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
swebench==3.0.15
sympy==1.13.3
synchronicity==0.9.12
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tqdm==4.67.1
transformers==4.49.0
typer==0.16.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
unidiff==0.7.5
urllib3==2.4.0
venn==0.1.3
virtualenv==20.31.2
watchfiles==1.0.5
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0 